refactor: separate model/ui state from ViewModel files
- Extract AuthUiState from AuthViewModel.kt into AuthUiState.kt
- Extract HomeFilmsData from HomeViewModel.kt into HomeFilmsData.kt
- Both remain in the same package so no import changes needed
