+++ app/src/main/java/com/drs/auralife/presentation/auth/AuthUiState.kt
@@ -1,0 +1,8 @@
+package com.drs.auralife.presentation.auth
+
+sealed class AuthUiState {
+    data object Idle : AuthUiState()
+    data object Loading : AuthUiState()
+    data class Success(val message: String) : AuthUiState()
+    data class Error(val message: String) : AuthUiState()
+}
