+++ app/src/main/java/com/drs/auralife/presentation/home/HomeFilmsData.kt
@@ -1,0 +1,8 @@
+package com.drs.auralife.presentation.home
+
+import com.drs.auralife.domain.model.Film
+
+data class HomeFilmsData(
+    val films: List<Film>,
+    val totalPages: Int,
+)
